feat: mustache middleware supports Module Worker mode
